requirements.txt
aiobreaker==1.2.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
tenacity==8.2.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
